backend/src/requirements.txt
# Requirements for deploying these functions to Google Cloud Functions
google-cloud-firestore
Flask  # only for local testing (the request object is flask.Request)
